+++ drivers/clk/at91/clk-usb.c
@@ -117,7 +117,7 @@ static int at91sam9x5_clk_usb_set_parent(struct clk_hw *hw, u8 index)
 	return 0;
 }
 
-static u8 at91sam9x5_clk_usb_get_parent(struct clk_hw *hw)
+static int at91sam9x5_clk_usb_get_parent(struct clk_hw *hw)
 {
 	struct at91sam9x5_clk_usb *usb = to_at91sam9x5_clk_usb(hw);
 	unsigned int usbr;
