+++ drivers/clk/socfpga/clk-pll.c
@@ -65,7 +65,7 @@ static unsigned long clk_pll_recalc_rate(struct clk_hw *hwclk,
 	return (unsigned long)vco_freq;
 }
 
-static u8 clk_pll_get_parent(struct clk_hw *hwclk)
+static int clk_pll_get_parent(struct clk_hw *hwclk)
 {
 	u32 pll_src;
 	struct socfpga_pll *socfpgaclk = to_socfpga_clk(hwclk);
